Moving drawing stuff to Drawable class (GameClient-owned equivalent of GameObject)
